As a side effect of trying to do things concurrently, code was measured and it became evident that it can run faster after small reorganization: polling at higher frequency, starting things in parallel or after removing superfluous delays.

* baseline: 34.84 seconds
* with these changes: 31.24 seconds
* with these changes and OCC changes (part of #94): 27.22 seconds

The slowest part currently  is building MVPD (13 seconds, 9 for building MVPD of the second CPU), reading cached version from PNOR should take just 50 milliseconds.